Initial upload of diff.sh, a proof of concept
Still very much in the planning stages here. My problem is (or rather, might become) one of my favorite backup tools urBackup not liking when files appear and disappear on it. Fair enough, but when this happens and it doesn't have a 'before copying' hash to compare with this new file's 'after copying' hash, it deletes the partial backup of it and every successfully backed up file this session, and will try again soon. So far this hasn't been a big problem in my test environment of a few hundred megabytes of data in use, but I imagine it will be a much bigger problem when I tell it to back up a terabyte or two of data in use. All of that creation and destruction happening while urBackup is trying to verify it copied something correctly will likely drive it mad. 

My goal with this snipit is to outsource the verification onto a program that won't give up in despair at every disparity. If this does become the problem I assume it will, I want to tell urBackup to stop checking its work, and just copy what it sees. Then I'll find a way to stick this string in the backup pipeline, checking the source and destination with md5, and emailing me the result of each backup. Then we'll rely on old fashioned Human Intuition to decide if the differing files are a problem or not. Hopefully I can script away the need for Human Intuition by the 4th major version, but we'll see how it goes until then.

Telling urBackup to ignore temporary files, like the MS Office files that start with "~$" has helped, but there are still legit, non-temporary files being created and deleted on this office's NAS, and we can't give up on this 'snapshot' of our data because urBackup wants a sterile working environment. Because it's a simple "appliance" style NAS, I don't think I could programatically take advantage of a disk quiesce from the other side of a Samba share, so this is the only solution I have thought of so far.