PowerShell script for measuring all home holders
Been thinking about migrating data, and wanted to get some info from a bunch of machines. I put the names in a text file, one per line, and fed it into this. Line 4 (where most of the work happens) I stole from someone else, and now can't remember. I'm a jerk.

For some reason, sometimes the Tee-Object outputs to screen and not to file, so I added line 6 as a backup.